separate input builder into individual modules
